if the configuration as shown <fileNamePattern >/srv/srv1.%d {HH}.%i.log </fileNamePattern > or <fileNamePattern >/srv/srv1.%d{HH} .log </fileNamePattern > the first file Will be deleted.
14:05:29,003 |-INFO in c.q.l.co.rolling.helper.RenameUtil - Renaming file [\apps\logs\log1\srv\srv.log] to [\apps\logs\log1\srv\srv.13.log]
14:05:29,004 |-INFO in c.q.l.core.rolling.helper.TimeBasedArchiveRemover - first clean up after appender initialization
14:05:29,004 |-INFO in c.q.l.core.rolling.helper.TimeBasedArchiveRemover - periodsElapsed = 336
14:05:29,006 |-INFO in c.q.l.core.rolling.helper.TimeBasedArchiveRemover - deleting \apps\logs\log1\srv\srv.13.log

+++ logback-core/src/main/java/ch/qos/logback/core/rolling/helper/SizeAndTimeBasedArchiveRemover.java
@@ -39,7 +39,11 @@ public class SizeAndTimeBasedArchiveRemover extends DefaultArchiveRemover {
         parentDir, stemRegex);
 
     for (File f : matchingFileArray) {
-      f.delete();
+      Date fileLastModified = rc.getRelativeDate(new Date(f.lastModified()), -1);
+    	if (fileLastModified.compareTo(dateOfPeriodToClean) <= 0){
+    		  f.delete();
+    	    addInfo("deleting " + f);
+    	}
     }
 
     if (parentClean) {
